+++ test/index.test.mjs
@@ -103,6 +103,30 @@ function testError()
 	{
 		expect(() =>
 		{
+			const constraints = {};
+			const input = 0;
+
+			adjuster.adjust(input, constraints);
+		}).toThrow(adjuster.CAUSE.NOT_OBJECT); // input must be an object
+
+		expect(() =>
+		{
+			const constraints = {};
+			const input = null;
+
+			adjuster.adjust(input, constraints);
+		}).toThrow(adjuster.CAUSE.NOT_OBJECT); // input must be an object; typeof null === "object"
+
+		expect(() =>
+		{
+			const constraints = {};
+			const input = [];
+
+			adjuster.adjust(input, constraints);
+		}).toThrow(adjuster.CAUSE.NOT_OBJECT); // input must be an object; typeof [] === "object"
+
+		expect(() =>
+		{
 			const constraints = {
 				id: adjuster.number().minValue(1),
 				name: adjuster.string().maxLength(16, true),