+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.14.0] - 2018-08-11
+### Added
+* Introduction in `README.md`
+
+### Changed
+* cause an error `adjuster.CAUSE.NOT_OBJECT` when `input` type of `adjuster.adjust()` is not an object
+
+### Fixed
+* build error in Windows
+
+### Others
+* support ES Modules (`*.mjs`)
+
 ## [0.13.0] - 2018-07-20
 ### Added
 * `adjuster.numberArray().eachInteger()`
@@ -138,7 +151,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.0] - 2018-04-18
 * First release.
 
-[Unreleased]: https://github.com/shimataro/node-adjuster/compare/v0.13.0...HEAD
+[Unreleased]: https://github.com/shimataro/node-adjuster/compare/v0.14.0...HEAD
+[0.14.0]: https://github.com/shimataro/node-adjuster/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/shimataro/node-adjuster/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/shimataro/node-adjuster/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/shimataro/node-adjuster/compare/v0.10.1...v0.11.0